chore(animated-header): Handling of loading and disabled states (#588)
* chore(animated-header): refactor

* chore(animated-header): loading/disabled states

* chore(animated-header): fixed dropdown onchange

* chore(animated-header): cleanup

* chore(animated-header): fix disabled state

* chore(animated-header): fix md layout overflow

* chore(animate-header): add license

* chore(animated-header): update tests

* chore(animated-header): fix md breakpoint layout

* chore(animated-header): fix content overflow

* chore(animated-header): remove console logs

* chore(animated-header): fix height animation

* chore(animated-header): fix skeletons safari

* chore(animated-header): format

* chore(animated-header): pr comments

* Update glass-tile.scss

As discussed with @motionmike let's change the skeleton loading colors to use $layer-01 for the background and $layer-02 for the element.

* Update ai-prompt-tile.scss

As discussed with Mike Olasov let's change the skeleton loading colors to use $layer-01 for the background and $layer-02 for the element.

* Update animated-header.scss

As discussed with Mike Olasov let's change the skeleton loading colors to use $layer-01 for the background and $layer-02 for the element.

* Update AIPromptTile.tsx

Fix classname edit

* Remove page load tile animation for Safari

Remove page load tile animation for Safari

* chore(animated-header): format

---------

Co-authored-by: Drew Glapa <Drew.Glapa@ibm.com>